requirements.txt
django-haystack==2.8.1
Django==2.2.13
elasticsearch==2.4.1
psycopg2-binary==2.8.4